requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
dash_daq==0.5.0
discord.py==2.3.2
numpy==1.26.4
pandas==2.1.1
pytest==7.4.1
ruff==0.5.0
# table2ascii
